Add guests to new piece of state on submit
